Implement modified authentication feature
Deleted scenario 'login user' from feature authentication to shorten
the waiting for the tests to run. It would be too much time to
implement the scenario considering the secure data handling and until
then this action saves time.

It would be also too time-consuming to manage to run scenarios
individually.